requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker==36.2.2
pip==25.3
pytest==8.3.5
selenium==4.29.0
selene==2.0.0rc9
flake8==7.3.0
black==25.11.0
isort==7.0.0
requests==2.32.5
jsonschema==4.25.1